Fetch env from SSM if not set
